+++ importlib_metadata/__init__.py
@@ -1,5 +1,5 @@
 from .api import distribution, Distribution, PackageNotFoundError  # noqa: F401
-from .api import metadata, entry_points, version, read_text, files
+from .api import metadata, entry_points, version, read_text, files, requires
 
 # Import for installation side-effects.
 from . import _hooks  # noqa: F401
@@ -10,6 +10,7 @@ __all__ = [
     'files',
     'metadata',
     'read_text',
+    'requires',
     'version',
     ]